requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
email-validator==1.3.1
pytest==7.4.2
Flask-Migrate==4.0.4
psycopg2-binary==2.9.10
